requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.16.0
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
numpy==2.3.1
openai==1.93.0
portalocker==2.10.1
protobuf==6.31.1
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
qdrant-client==1.14.3
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
